fix: errors with config

+++ src/schemas/movies.py
@@ -18,7 +18,7 @@ class MovieDetailResponseSchema(BaseModel):
     orig_title: str
     status: str
     orig_lang: str
-    budget: Decimal
+    budget: float
     revenue: float
     country: str
